Abstracted EventDashboard, added Bottom Nav buttons to Dashboard
