+++ Spawn-Android/app/src/main/java/com/example/spawn_app_android/ui/navigation/bottomNavigation.kt
@@ -27,6 +27,9 @@ fun BottomNavigation(
 ) {
     NavigationBar(
         modifier = Modifier.fillMaxWidth(),
+        containerColor = Color.Transparent,
+        contentColor = Color.Transparent,
+        tonalElevation = 0.dp
     ) {
         Row(
             modifier = Modifier.fillMaxWidth(),
